@@ -1,8 +1,9 @@
 <script>
 	import { getContext } from 'svelte';
 	import GeolocationIcon from './GeolocationIcon.svelte';
+	// import { FOCUSABLE } from '../contexts.js';
 
-	const tabindex = getContext('focusable');
+	const tabindex = getContext('FOCUSABLE');
 </script>
 
 <div>
